refactor: split shared-service tenancy analysis
