+++ implementations/python/packages/aces_sdl/semantics/deployment_tenancy.py
@@ -4,50 +4,21 @@ from __future__ import annotations
 
 from collections import defaultdict
 from collections.abc import Callable, Mapping
-from dataclasses import dataclass
 
-from ..deployment_tenancy import (
-    EndpointPersona,
-    StateOwner,
-    TenantIsolationMode,
-    WorkloadAuthenticationMode,
-)
+from ..deployment_tenancy import EndpointPersona
 from ..nodes import NodeType
 from ..relationships import RelationshipType
+from ._deployment_shared_services import (
+    cross_cell_service_consumption_issues,
+    shared_service_issues,
+)
+from ._deployment_tenancy_types import (
+    CellIndex,
+    DeploymentTenancyIssue,
+    enum_value,
+    issue,
+)
 from ._domain_topology_types import resolve_section_ref
-
-
-@dataclass(frozen=True)
-class DeploymentTenancyIssue:
-    code: str
-    message: str
-
-
-def _issue(code: str, message: str) -> DeploymentTenancyIssue:
-    return DeploymentTenancyIssue(code=code, message=message)
-
-
-def _enum_value(value: object) -> str:
-    return getattr(value, "value", str(value))
-
-
-def _service_owner(service_ref: object, nodes: Mapping[str, object]) -> str | None:
-    if not isinstance(service_ref, str):
-        return None
-    for node_name, node in nodes.items():
-        for service in getattr(node, "services", ()):
-            name = getattr(service, "name", "")
-            if name and service_ref == f"nodes.{node_name}.services.{name}":
-                return node_name
-    return None
-
-
-@dataclass
-class _CellIndex:
-    node_cell: dict[str, str]
-    cell_tenant: dict[str, str]
-    tenant_nodes: defaultdict[str, set[str]]
-    issues: list[DeploymentTenancyIssue]
 
 
 def _index_cell_members(
@@ -55,7 +26,7 @@ def _index_cell_members(
     cell: object,
     tenant_name: str | None,
     nodes: Mapping[str, object],
-    index: _CellIndex,
+    index: CellIndex,
     is_unresolved: Callable[[object], bool],
 ) -> None:
     for node_ref in getattr(cell, "node_refs", ()):
@@ -64,7 +35,7 @@ def _index_cell_members(
         node_name = resolve_section_ref(node_ref, "nodes", nodes)
         if node_name is None:
             index.issues.append(
-                _issue(
+                issue(
                     "deployment-tenancy.cell.node-unbound",
                     f"Deployment cell '{cell_name}' node_ref '{node_ref}' does not resolve to a node",
                 )
@@ -73,7 +44,7 @@ def _index_cell_members(
         previous = index.node_cell.get(node_name)
         if previous is not None and previous != cell_name:
             index.issues.append(
-                _issue(
+                issue(
                     "deployment-tenancy.cell.node-multiple",
                     f"Node '{node_name}' belongs to multiple deployment cells: {previous}, {cell_name}",
                 )
@@ -92,7 +63,7 @@ def _missing_vm_issues(
     if not deployment_cells:
         return []
     return [
-        _issue(
+        issue(
             "deployment-tenancy.cell.node-missing",
             f"VM node '{node_name}' must belong to exactly one deployment cell",
         )
@@ -106,8 +77,8 @@ def _build_cell_index(
     deployment_cells: Mapping[str, object],
     nodes: Mapping[str, object],
     is_unresolved: Callable[[object], bool],
-) -> _CellIndex:
-    index = _CellIndex({}, {}, defaultdict(set), [])
+) -> CellIndex:
+    index = CellIndex({}, {}, defaultdict(set), [])
     for cell_name, cell in deployment_cells.items():
         tenant_ref = getattr(cell, "tenant_ref", "")
         tenant_name = None
@@ -115,7 +86,7 @@ def _build_cell_index(
             tenant_name = resolve_section_ref(tenant_ref, "deployment_tenants", deployment_tenants)
             if tenant_name is None:
                 index.issues.append(
-                    _issue(
+                    issue(
                         "deployment-tenancy.cell.tenant-unbound",
                         f"Deployment cell '{cell_name}' tenant_ref '{tenant_ref}' does not resolve "
                         "to a deployment tenant",
@@ -144,12 +115,12 @@ def _relationship_detail_issues(
         type_value = getattr(relationship, "type", "")
         if is_unresolved(type_value):
             continue
-        type_name = _enum_value(type_value)
+        type_name = enum_value(type_value)
         expected = _DETAIL_FIELDS.get(type_name)
         populated = [field_name for field_name in detail_fields if getattr(relationship, field_name, None) is not None]
         if expected is not None and expected not in populated:
             issues.append(
-                _issue(
+                issue(
                     "deployment-tenancy.relationship.detail-required",
                     f"Relationship '{name}' type '{type_name}' requires {expected} detail",
                 )
@@ -157,7 +128,7 @@ def _relationship_detail_issues(
         for field_name in populated:
             if field_name != expected:
                 issues.append(
-                    _issue(
+                    issue(
                         "deployment-tenancy.relationship.detail-mismatch",
                         f"Relationship '{name}' carries {field_name} detail with type '{type_name}'",
                     )
@@ -168,14 +139,14 @@ def _relationship_detail_issues(
 def _placement_issues(
     nodes: Mapping[str, object],
     relationships: Mapping[str, object],
-    cell_index: _CellIndex,
+    cell_index: CellIndex,
     is_unresolved: Callable[[object], bool],
 ) -> list[DeploymentTenancyIssue]:
     issues: list[DeploymentTenancyIssue] = []
     carrier_by_source: dict[str, str] = {}
     edge_name_by_source: dict[str, str] = {}
     for name, relationship in relationships.items():
-        if _enum_value(getattr(relationship, "type", "")) != RelationshipType.PLACED_ON_CARRIER.value:
+        if enum_value(getattr(relationship, "type", "")) != RelationshipType.PLACED_ON_CARRIER.value:
             continue
         if getattr(relationship, "carrier_placement", None) is None:
             continue
@@ -209,7 +180,7 @@ def _placement_endpoint_issue(
         or getattr(nodes.get(source), "type", None) != NodeType.VM
         or getattr(nodes.get(target), "type", None) != NodeType.VM
     ):
-        return _issue(
+        return issue(
             "deployment-tenancy.placement.endpoint-invalid",
             f"Relationship '{relationship_name}' carrier placement endpoints must resolve to VM nodes",
         )
@@ -221,13 +192,13 @@ def _placement_binding_issues(
     source: str,
     target: str,
     nodes: Mapping[str, object],
-    cell_index: _CellIndex,
+    cell_index: CellIndex,
     carrier_by_source: Mapping[str, str],
 ) -> list[DeploymentTenancyIssue]:
     issues: list[DeploymentTenancyIssue] = []
     if source == target:
         issues.append(
-            _issue(
+            issue(
                 "deployment-tenancy.placement.self",
                 f"Relationship '{relationship_name}' cannot place a node on itself",
             )
@@ -235,15 +206,15 @@ def _placement_binding_issues(
     previous = carrier_by_source.get(source)
     if previous is not None and previous != target:
         issues.append(
-            _issue(
+            issue(
                 "deployment-tenancy.placement.multiple-carriers",
                 f"Node '{source}' has multiple carrier placements",
             )
         )
-    persona = _enum_value(getattr(nodes[target], "endpoint_persona", ""))
+    persona = enum_value(getattr(nodes[target], "endpoint_persona", ""))
     if persona != EndpointPersona.CARRIER.value:
         issues.append(
-            _issue(
+            issue(
                 "deployment-tenancy.placement.target-not-carrier",
                 f"Relationship '{relationship_name}' target '{target}' must have endpoint persona 'carrier'",
             )
@@ -252,7 +223,7 @@ def _placement_binding_issues(
     target_cell = cell_index.node_cell.get(target)
     if source_cell is not None and target_cell is not None and source_cell != target_cell:
         issues.append(
-            _issue(
+            issue(
                 "deployment-tenancy.placement.cross-cell",
                 f"Relationship '{relationship_name}' carrier placement crosses different deployment cells",
             )
@@ -276,7 +247,7 @@ def _placement_graph_issues(
     edge_name_by_source: Mapping[str, str],
 ) -> list[DeploymentTenancyIssue]:
     issues = [
-        _issue(
+        issue(
             "deployment-tenancy.placement.cycle",
             f"Relationship '{edge_name_by_source[start]}' participates in a carrier placement cycle",
         )
@@ -284,307 +255,13 @@ def _placement_graph_issues(
         if _has_placement_cycle(start, carrier_by_source)
     ]
     issues.extend(
-        _issue(
+        issue(
             "deployment-tenancy.placement.nested",
             f"Relationship '{edge_name_by_source[source]}' cannot place a node on another placed node",
         )
         for source, target in carrier_by_source.items()
         if target in carrier_by_source
     )
-    return issues
-
-
-def _state_consumer_nodes(volume: object) -> set[str]:
-    return {getattr(consumer, "node", "") for consumer in getattr(volume, "consumers", ())}
-
-
-@dataclass(frozen=True)
-class _SharedServiceContext:
-    relationship_name: str
-    tenant_name: str
-    service_node: str
-    service_tenant: str
-    isolation: str
-    authentication: str
-    state_owner: str
-    reset_owner: str
-    state_refs: tuple[object, ...]
-
-
-def _shared_service_context(
-    relationship_name: str,
-    relationship: object,
-    deployment_tenants: Mapping[str, object],
-    nodes: Mapping[str, object],
-    cell_index: _CellIndex,
-    is_unresolved: Callable[[object], bool],
-) -> tuple[_SharedServiceContext | None, DeploymentTenancyIssue | None]:
-    detail = getattr(relationship, "shared_service", None)
-    if detail is None:
-        return None, None
-    source_ref = getattr(relationship, "source", "")
-    target_ref = getattr(relationship, "target", "")
-    if any(is_unresolved(value) for value in (source_ref, target_ref)):
-        return None, None
-    tenant_name = resolve_section_ref(source_ref, "deployment_tenants", deployment_tenants)
-    service_node = _service_owner(target_ref, nodes)
-    if tenant_name is None or service_node is None:
-        return None, _issue(
-            "deployment-tenancy.shared-service.endpoint-invalid",
-            f"Relationship '{relationship_name}' must connect a deployment tenant to a named VM service",
-        )
-    service_cell = cell_index.node_cell.get(service_node)
-    service_tenant = cell_index.cell_tenant.get(service_cell, "") if service_cell is not None else ""
-    return (
-        _SharedServiceContext(
-            relationship_name=relationship_name,
-            tenant_name=tenant_name,
-            service_node=service_node,
-            service_tenant=service_tenant,
-            isolation=_enum_value(getattr(detail, "tenant_isolation", "")),
-            authentication=_enum_value(getattr(detail, "workload_authentication", "")),
-            state_owner=_enum_value(getattr(detail, "mutable_state_owner", "")),
-            reset_owner=_enum_value(getattr(detail, "reset_generation_owner", "")),
-            state_refs=tuple(getattr(detail, "mutable_state_refs", ())),
-        ),
-        None,
-    )
-
-
-def _shared_service_policy_issues(context: _SharedServiceContext) -> list[DeploymentTenancyIssue]:
-    issues: list[DeploymentTenancyIssue] = []
-    if context.tenant_name != context.service_tenant and (
-        context.isolation not in {TenantIsolationMode.STATELESS.value, TenantIsolationMode.TENANT_PARTITIONED.value}
-        or context.authentication != WorkloadAuthenticationMode.TENANT_SCOPED_WORKLOAD_IDENTITY.value
-    ):
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.cross-tenant-unsafe",
-                f"Relationship '{context.relationship_name}' {context.isolation or 'cross-tenant'} isolation "
-                "requires tenant-scoped workload authentication",
-            )
-        )
-    if (
-        context.isolation == TenantIsolationMode.STATELESS.value
-        and context.state_owner == StateOwner.SHARED_SERVICE.value
-    ):
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.stateless-service-state",
-                f"Relationship '{context.relationship_name}' stateless isolation forbids "
-                "shared-service-owned mutable state",
-            )
-        )
-    if context.isolation == TenantIsolationMode.TENANT_PARTITIONED.value and (
-        not context.state_refs
-        or context.state_owner != StateOwner.SHARED_SERVICE.value
-        or context.reset_owner != StateOwner.SHARED_SERVICE.value
-    ):
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.partitioned-state-required",
-                f"Relationship '{context.relationship_name}' tenant_partitioned isolation requires "
-                "shared-service-owned state and reset generation",
-            )
-        )
-    if context.state_refs and context.state_owner == StateOwner.NONE.value:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.state-owner-missing",
-                f"Relationship '{context.relationship_name}' mutable_state_refs require a non-none owner",
-            )
-        )
-    if not context.state_refs and context.state_owner != StateOwner.NONE.value:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.state-refs-missing",
-                f"Relationship '{context.relationship_name}' mutable state owner requires mutable_state_refs",
-            )
-        )
-    if context.state_owner != StateOwner.NONE.value and context.reset_owner != context.state_owner:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.reset-owner-mismatch",
-                f"Relationship '{context.relationship_name}' reset_generation_owner must equal mutable state owner",
-            )
-        )
-    if context.state_owner == StateOwner.NONE.value and context.reset_owner != StateOwner.NONE.value:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.reset-owner-without-state",
-                f"Relationship '{context.relationship_name}' reset_generation_owner must be none without mutable state",
-            )
-        )
-    return issues
-
-
-def _shared_state_ref_issues(
-    state_ref: object,
-    context: _SharedServiceContext,
-    persistent_volumes: Mapping[str, object],
-    cell_index: _CellIndex,
-    state_owners: dict[str, tuple[str, str]],
-    is_unresolved: Callable[[object], bool],
-) -> list[DeploymentTenancyIssue]:
-    if is_unresolved(state_ref):
-        return []
-    state_name = resolve_section_ref(state_ref, "persistent_volumes", persistent_volumes)
-    if state_name is None:
-        return [
-            _issue(
-                "deployment-tenancy.shared-service.state-unbound",
-                f"Relationship '{context.relationship_name}' mutable_state_ref '{state_ref}' does not resolve "
-                "to a persistent volume",
-            )
-        ]
-    issues: list[DeploymentTenancyIssue] = []
-    owner_key = (context.tenant_name, context.state_owner)
-    previous = state_owners.setdefault(state_name, owner_key)
-    if previous != owner_key:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.state-owner-conflict",
-                f"Persistent volume '{state_name}' has conflicting shared-service owners",
-            )
-        )
-    consumers = _state_consumer_nodes(persistent_volumes[state_name])
-    if context.state_owner == StateOwner.CONSUMER_TENANT.value:
-        allowed = cell_index.tenant_nodes.get(context.tenant_name, set())
-        if not consumers or not consumers.issubset(allowed):
-            issues.append(
-                _issue(
-                    "deployment-tenancy.shared-service.consumer-state-mismatch",
-                    f"Relationship '{context.relationship_name}' consumer-owned state must be consumed "
-                    "only by its tenant",
-                )
-            )
-    elif context.state_owner == StateOwner.SHARED_SERVICE.value and context.service_node not in consumers:
-        issues.append(
-            _issue(
-                "deployment-tenancy.shared-service.service-state-mismatch",
-                f"Relationship '{context.relationship_name}' shared-service-owned state must be consumed "
-                "by the service node",
-            )
-        )
-    return issues
-
-
-def _shared_service_issues(
-    deployment_tenants: Mapping[str, object],
-    nodes: Mapping[str, object],
-    persistent_volumes: Mapping[str, object],
-    relationships: Mapping[str, object],
-    cell_index: _CellIndex,
-    is_unresolved: Callable[[object], bool],
-) -> list[DeploymentTenancyIssue]:
-    issues: list[DeploymentTenancyIssue] = []
-    state_owners: dict[str, tuple[str, str]] = {}
-    for name, relationship in relationships.items():
-        if _enum_value(getattr(relationship, "type", "")) != RelationshipType.USES_SHARED_SERVICE.value:
-            continue
-        context, endpoint_issue = _shared_service_context(
-            name,
-            relationship,
-            deployment_tenants,
-            nodes,
-            cell_index,
-            is_unresolved,
-        )
-        if endpoint_issue is not None:
-            issues.append(endpoint_issue)
-            continue
-        if context is None:
-            continue
-        issues.extend(_shared_service_policy_issues(context))
-        for state_ref in context.state_refs:
-            issues.extend(
-                _shared_state_ref_issues(
-                    state_ref,
-                    context,
-                    persistent_volumes,
-                    cell_index,
-                    state_owners,
-                    is_unresolved,
-                )
-            )
-    return issues
-
-
-def _shared_service_permissions(
-    deployment_tenants: Mapping[str, object],
-    nodes: Mapping[str, object],
-    relationships: Mapping[str, object],
-    is_unresolved: Callable[[object], bool],
-) -> set[tuple[str, str]]:
-    permitted: set[tuple[str, str]] = set()
-    for relationship in relationships.values():
-        if _enum_value(getattr(relationship, "type", "")) != RelationshipType.USES_SHARED_SERVICE.value:
-            continue
-        source_ref = getattr(relationship, "source", "")
-        target_ref = getattr(relationship, "target", "")
-        if any(is_unresolved(value) for value in (source_ref, target_ref)):
-            continue
-        tenant_name = resolve_section_ref(source_ref, "deployment_tenants", deployment_tenants)
-        if tenant_name is not None and _service_owner(target_ref, nodes) is not None:
-            permitted.add((tenant_name, target_ref))
-    return permitted
-
-
-def _cross_cell_consumption_issue(
-    relationship_name: str,
-    relationship: object,
-    nodes: Mapping[str, object],
-    cell_index: _CellIndex,
-    permitted: set[tuple[str, str]],
-    is_unresolved: Callable[[object], bool],
-) -> DeploymentTenancyIssue | None:
-    if _enum_value(getattr(relationship, "type", "")) == RelationshipType.USES_SHARED_SERVICE.value:
-        return None
-    source_ref = getattr(relationship, "source", "")
-    target_ref = getattr(relationship, "target", "")
-    if any(is_unresolved(value) for value in (source_ref, target_ref)):
-        return None
-    source_node = resolve_section_ref(source_ref, "nodes", nodes) or _service_owner(source_ref, nodes)
-    target_node = _service_owner(target_ref, nodes)
-    if source_node is None or target_node is None:
-        return None
-    source_cell = cell_index.node_cell.get(source_node)
-    target_cell = cell_index.node_cell.get(target_node)
-    if source_cell is None or target_cell is None or source_cell == target_cell:
-        return None
-    source_tenant = cell_index.cell_tenant.get(source_cell)
-    if source_tenant is None or (source_tenant, target_ref) in permitted:
-        return None
-    return _issue(
-        "deployment-tenancy.shared-service.binding-required",
-        f"Relationship '{relationship_name}' cross-cell service consumption requires an explicit "
-        "shared-service binding for the consumer tenant and target service",
-    )
-
-
-def _cross_cell_service_consumption_issues(
-    deployment_tenants: Mapping[str, object],
-    nodes: Mapping[str, object],
-    relationships: Mapping[str, object],
-    cell_index: _CellIndex,
-    is_unresolved: Callable[[object], bool],
-) -> list[DeploymentTenancyIssue]:
-    if not cell_index.node_cell:
-        return []
-
-    permitted = _shared_service_permissions(deployment_tenants, nodes, relationships, is_unresolved)
-    issues: list[DeploymentTenancyIssue] = []
-    for name, relationship in relationships.items():
-        issue = _cross_cell_consumption_issue(
-            name,
-            relationship,
-            nodes,
-            cell_index,
-            permitted,
-            is_unresolved,
-        )
-        if issue is not None:
-            issues.append(issue)
     return issues
 
 
@@ -604,7 +281,7 @@ def analyze_deployment_tenancy(
     issues.extend(_relationship_detail_issues(relationships, is_unresolved))
     issues.extend(_placement_issues(nodes, relationships, cell_index, is_unresolved))
     issues.extend(
-        _shared_service_issues(
+        shared_service_issues(
             deployment_tenants,
             nodes,
             persistent_volumes,
@@ -614,7 +291,7 @@ def analyze_deployment_tenancy(
         )
     )
     issues.extend(
-        _cross_cell_service_consumption_issues(
+        cross_cell_service_consumption_issues(
             deployment_tenants,
             nodes,
             relationships,
